add common cert utils.
